third_party/blink/web_tests/platform/generic/http/tests/images/document-policy-oversized-images-expected.txt

CONSOLE ERROR: Document policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Document policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Document policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Document policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Document policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Document policy violation: oversized-images is not allowed in this document.
